requirements.txt
certifi
chardet
defusedxml
Django => 2.1, < 2.2
django-countries
django-allauth
django-mock-queries
django-rest-auth
djangorestframework
freezegun
idna
oauthlib
psycopg2
pytest
pytest-cov
pytest-django
python-dateutil
python3-openid
pytz
PyYAML
requests
requests-oauthlib
six
urllib3
